requirements.txt
streamlit
langchain
langchain_openai
langchain_community
chromadb
python-dotenv
tiktoken